Changed native interface to comply with sys_treetank.
git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@4383 e3ddb328-5bfe-0310-b762-aafcbcbd2528

+++ src/main/openbsd/NativeTreeTank.c
@@ -28,7 +28,7 @@
 #include <stdio.h>
 #include <string.h>
 
-JNIEXPORT jshort JNICALL Java_org_treetank_pagelayer_CryptoNativeImpl_syscall(
+JNIEXPORT jint JNICALL Java_org_treetank_pagelayer_CryptoNativeImpl_syscall(
   JNIEnv *env,
   jobject o,
   jbyte tank,
@@ -36,10 +36,10 @@ JNIEXPORT jshort JNICALL Java_org_treetank_pagelayer_CryptoNativeImpl_syscall(
   jshort length,
   jobject buffer)
 {
-  jshort result    = 0x0;
-  u_int8_t *bufferPtr = (*env)->GetDirectBufferAddress(env, buffer);
+  jint   result    = 0x0;
+  jbyte *bufferPtr = (*env)->GetDirectBufferAddress(env, buffer);
   
-  result = (jshort) syscall(306, tank, operation, length, bufferPtr);
+  result = syscall(306, tank, operation, length, bufferPtr);
 
   return result;
 }
